+++ messages/src/main/java/com/jwoglom/pumpx2/pump/messages/models/InsulinUnit.java
@@ -2,8 +2,8 @@ package com.jwoglom.pumpx2.pump.messages.models;
 
 public class InsulinUnit {
     // 1000 = 1 unit
-    public static float from1000To1(Long units) {
-        return (units.floatValue() / 1000);
+    public static double from1000To1(Long units) {
+        return (units.doubleValue() / 1000);
     }
 
     public static long from1To1000(Float units) {
